+++ CHANGELOG.md
@@ -5,13 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased : 1.0]
+
+- inotify : update fluxboxlaucher config when external config file was edit.
+
+## [Unreleased : 0.3]
+
+- add a command line editor and test it
+- search name and icon (data in .desktop file) corresponding
 
 ## [Unreleased : 0.2]
 ### Added
 
-- add a command line editor and test it
-- search name and icon (data in .desktop file) corresponding
 - unit tests
+- refactoring
 
 ## [0.1]
 ### Added