+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased : 0.2]
 ### Added
 
+- i18n on Nix without patching
 - unit tests
 - refactoring