+++ CHANGELOG.md
@@ -13,8 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - add a command line editor and test it
 - search name and icon (data in .desktop file) corresponding
+- unit tests on locales and gettext
 
-## [Unreleased : 0.2]
+## [0.2]
 ### Added
 
 - i18n on Nix without patching